+++ CHANGELOG.md
@@ -53,6 +53,7 @@ Changelog
  * moved API interfaces to `Api` sub-namespace
  * removed notions of "directories" and "files". All resources can have children
    and a body now.
+ * renamed `Selector` to `Glob`
 
 * 1.0.0-alpha4 (2014-12-03)